Move visualize into thresholds_tools
